Add files via upload
This Excel file includes the top 1000 YouTube channels and information about them. I have created a VLOOKUP function to search for certain ranks and the specific YouTube channel. I have also created a pivot table to show the different categories of YouTube channels and the sum of subscribers for those categories. There is also a filter to break down which channels were created in each year. A pie chart is included for easy visualization.